2010-06-04  Ramon Novoa  <rnovoa@artica.es>
	* pandora_console.spec: Fixed. Upgrading an RPM broke the installation.




git-svn-id: https://svn.code.sf.net/p/pandora/code/trunk@2856 c3f86ba8-e40f-0410-aaad-9ba5e7f4b01f

+++ pandora_console/pandora_console.spec
@@ -70,6 +70,12 @@ fi
 
 
 %preun
+
+# Upgrading
+if [ "$1" = "1" ]; then
+        exit 0
+fi
+
 rm -Rf %{prefix}/pandora_console
 
 %files
